+++ packages/smarthr-ui/src/components/AppHeader/multilingualization/translate.ts
@@ -1,7 +1,0 @@
-import { type Messages, translation } from './messages'
-import { DEFAULT_LOCALE, type Locale } from './types'
-
-export const translate = <ID extends keyof Messages>(id: ID, locale?: Locale | null) =>
-  translation[locale ?? DEFAULT_LOCALE][
-    id
-  ] as (typeof translation)[typeof DEFAULT_LOCALE][typeof id]