+++ packages/smarthr-ui/src/components/Dialog/ActionDialog/stories/ActionDialog.stories.tsx
@@ -223,9 +223,16 @@ export const PortalParent: StoryObj<typeof ActionDialog> = {
 }
 
 export const Width: StoryObj<typeof ActionDialog> = {
-  name: 'width',
+  name: 'width（非推奨）',
   args: {
     width: _widthOptions.string,
+  },
+}
+
+export const Size: StoryObj<typeof ActionDialog> = {
+  name: 'size',
+  args: {
+    size: 'M',
   },
 }
 
@@ -246,7 +253,6 @@ export const FirstFocusTarget: StoryObj<typeof ActionDialog> = {
         >
           <label>
             入力要素
-            {/* eslint-disable-next-line smarthr/a11y-input-has-name-attribute, smarthr/a11y-input-in-form-control */}
             <Input ref={inputRef} />
           </label>
         </ActionDialog>