+++ packages/smarthr-ui/src/components/Dialog/stories/Dialog.stories.tsx
@@ -57,9 +57,16 @@ export const Id: StoryObj<typeof Dialog> = {
 }
 
 export const Width: StoryObj<typeof Dialog> = {
-  name: 'width',
+  name: 'width（非推奨）',
   args: {
     width: _widthOptions.string,
+  },
+}
+
+export const Size: StoryObj<typeof Dialog> = {
+  name: 'size',
+  args: {
+    size: 'M',
   },
 }
 
@@ -75,7 +82,6 @@ export const FirstFocusTarget: StoryObj<typeof Dialog> = {
         <Dialog {...args} firstFocusTarget={inputRef} isOpen={open} onPressEscape={handleClose}>
           <label>
             入力要素
-            {/* eslint-disable-next-line smarthr/a11y-input-has-name-attribute, smarthr/a11y-input-in-form-control */}
             <Input ref={inputRef} />
           </label>
         </Dialog>
@@ -91,7 +97,6 @@ export const IsOpen: StoryObj<typeof Dialog> = {
     const renderCheckbox = (
       <>
         <p>isOpen: {String(open)}</p>
-        {/* eslint-disable-next-line smarthr/a11y-input-has-name-attribute */}
         <Checkbox checked={open} onChange={() => setOpen(!open)}>
           isOpen
         </Checkbox>