+++ packages/smarthr-ui/src/components/Dialog/stories/DialogContent.stories.tsx
@@ -54,9 +54,16 @@ export const Id: StoryObj<typeof DialogContent> = {
 }
 
 export const Width: StoryObj<typeof DialogContent> = {
-  name: 'width',
+  name: 'width（非推奨）',
   args: {
     width: _widthOptions.number,
+  },
+}
+
+export const Size: StoryObj<typeof DialogContent> = {
+  name: 'size',
+  args: {
+    size: 'M',
   },
 }
 
@@ -72,7 +79,6 @@ export const FirstFocusTarget: StoryObj<typeof DialogContent> = {
         <DialogContent {...args} firstFocusTarget={inputRef}>
           <label>
             入力要素
-            {/* eslint-disable-next-line smarthr/a11y-input-has-name-attribute, smarthr/a11y-input-in-form-control */}
             <Input ref={inputRef} />
           </label>
           <DialogCloser>