+++ packages/smarthr-ui/src/components/Table/ThCheckbox.tsx
@@ -1,7 +1,10 @@
+'use client'
+
 import { type ComponentProps, forwardRef, useMemo } from 'react'
 import { tv } from 'tailwind-variants'
 
 import { type DecoratorsType, useDecorators } from '../../hooks/useDecorators'
+import { useIntl } from '../../intl'
 import { Balloon } from '../Balloon'
 import { Checkbox, type Props as CheckboxProps } from '../Checkbox'
 
@@ -12,13 +15,7 @@ type Props = {
   decorators?: DecoratorsType<'checkAllInvisibleLabel'> & {
     checkColumnName?: (text: string) => string
   }
-} & Pick<ComponentProps<typeof Th>, 'vAlign'>
-
-const DECORATOR_DEFAULT_TEXTS = {
-  checkAllInvisibleLabel: 'すべての項目を選択/解除',
-  checkColumnName: '選択',
-} as const
-type DecoratorKeyTypes = keyof typeof DECORATOR_DEFAULT_TEXTS
+} & Pick<ComponentProps<typeof Th>, 'vAlign' | 'fixed'>
 
 const classNameGenerator = tv({
   slots: {
@@ -39,7 +36,28 @@ const classNameGenerator = tv({
 })
 
 export const ThCheckbox = forwardRef<HTMLInputElement, CheckboxProps & Props>(
-  ({ vAlign, decorators, className, ...others }, ref) => {
+  ({ vAlign, fixed, decorators, className, ...others }, ref) => {
+    const { localize } = useIntl()
+
+    const decoratorDefaultTexts = useMemo(
+      () => ({
+        checkAllInvisibleLabel: localize({
+          id: 'smarthr-ui/ThCheckbox/checkAllInvisibleLabel',
+          defaultText: 'すべての項目を選択/解除',
+        }),
+        checkColumnName: localize({
+          id: 'smarthr-ui/ThCheckbox/checkColumnName',
+          defaultText: '選択',
+        }),
+      }),
+      [localize],
+    )
+
+    const decorated = useDecorators<'checkAllInvisibleLabel' | 'checkColumnName'>(
+      decoratorDefaultTexts,
+      decorators,
+    )
+
     const classNames = useMemo(() => {
       const { wrapper, inner, balloon, checkbox } = classNameGenerator()
 
@@ -51,12 +69,11 @@ export const ThCheckbox = forwardRef<HTMLInputElement, CheckboxProps & Props>(
       }
     }, [className])
 
-    const decorated = useDecorators<DecoratorKeyTypes>(DECORATOR_DEFAULT_TEXTS, decorators)
-
     return (
       // Th に必要な属性やイベントは不要
       <Th
         vAlign={vAlign}
+        fixed={fixed}
         className={classNames.wrapper}
         aria-label={decorated.checkColumnName as string}
       >