+++ packages/smarthr-ui/src/smarthr-ui-preset.ts
@@ -6,6 +6,7 @@ import { defaultColor } from './themes/createColor'
 import { defaultFontSize, defaultHtmlFontSize } from './themes/createFontSize'
 import { defaultShadow } from './themes/createShadow/defaultShadow'
 import { createSpacingByChar, primitiveTokens as spacingSizes } from './themes/createSpacing'
+import { defaultWidth } from './themes/createWidth'
 import { defaultZIndex } from './themes/createZIndex'
 
 import type { Config } from 'tailwindcss'
@@ -35,14 +36,14 @@ defaultConfig.twMergeConfig = {
       'border-b-shorthand',
       'border-l-shorthand',
     ],
-    'font-size': [
+    fontSize: [
       {
         text: ['2xs', 'xs', 'sm', 'base', 'lg', 'xl', '2xl', 'inherit'],
       },
     ],
     lineHeight: [
       {
-        leading: ['none', 'tight', 'normal', 'loose'],
+        leading: ['none', 'tight', 'normal', 'loose', '[0]'],
       },
     ],
     zIndex: [
@@ -54,7 +55,6 @@ defaultConfig.twMergeConfig = {
           'fixed-menu',
           'overlap-base',
           'overlap',
-          'flash-message',
           (classPart: string) => /^\[\d+\]$/.test(classPart),
         ],
       },
@@ -178,6 +178,7 @@ export default {
       loose: '1.75',
     },
     maxWidth: {
+      ...defaultWidth,
       none: 'none',
       full: '100%',
       min: 'min-content',
@@ -221,11 +222,13 @@ export default {
       'fixed-menu': `${defaultZIndex.FIXED_MENU}`,
       'overlap-base': `${defaultZIndex.OVERLAP_BASE}`,
       overlap: `${defaultZIndex.OVERLAP}`,
-      'flash-message': `${defaultZIndex.FLASH_MESSAGE}`,
     },
     extend: {
       aria: {
         'current-page': 'current="page"',
+      },
+      width: {
+        ...defaultWidth,
       },
       minHeight: ({ theme }) => ({
         ...theme('spacing'),
@@ -368,23 +371,6 @@ export default {
           to: {
             opacity: '1',
             transform: 'translateY(0)',
-          },
-        },
-        'flash-message-bounce': {
-          'from, 20%, 53%, 80%, to': {
-            'animation-timing-function': 'cubic-bezier(0.215, 0.61, 0.355, 1)',
-            transform: 'translate3d(0, 0, 0)',
-          },
-          '40%, 43%': {
-            'animation-timing-function': 'cubic-bezier(0.755, 0.05, 0.855, 0.06)',
-            transform: 'translate3d(0, -30px, 0)',
-          },
-          '70%': {
-            'animation-timing-function': 'cubic-bezier(0.755, 0.05, 0.855, 0.06)',
-            transform: 'translate3d(0, -15px, 0)',
-          },
-          '90%': {
-            transform: 'translate3d(0, -4px, 0)',
           },
         },
       }),